created test folder and moved all testing to that folder
