+++ tests/client/components/user-home.spec.js
@@ -4,7 +4,7 @@ import {expect} from 'chai'
 import React from 'react'
 import enzyme, {shallow} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import {UserHome} from './user-home'
+import {UserHome} from '../../../client/components/user-home'
 
 const adapter = new Adapter()
 enzyme.configure({adapter})
